+++ app/src/entities/IndicatorType.ts
@@ -1,0 +1,5 @@
+export enum INDICATOR_TYPE {
+  smiley = "smiley",
+  gauge = "gauge",
+  boolean = "boolean",
+}