+++ app/src/scenes/onboarding/screens.js
@@ -9,12 +9,12 @@ import IllustrationOnboarding22Svg from "../../../assets/svg/IllustrationOnboard
 import { colors } from "@/utils/colors";
 
 import { StyleSheet, View, Dimensions } from "react-native";
-export const buttonHeight = 43;
-export const buttonSmallHeight = 30;
-export const defaultPadding = Math.min(Dimensions.get("window").width * 0.7, 30);
+const buttonHeight = 43;
+const buttonSmallHeight = 30;
+const defaultPadding = Math.min(Dimensions.get("window").width * 0.7, 30);
 export const screenWidth = Dimensions.get("window").width;
-export const screenHeight = Dimensions.get("window").height;
-export const menuHeight = 80;
+const screenHeight = Dimensions.get("window").height;
+const menuHeight = 80;
 const size = screenWidth * 0.25;
 
 export const Screen0 = () => (