+++ app/src/utils/liste_indicateurs.1.ts
@@ -1,5 +1,7 @@
-import { INDICATOR_ORDER, INDICATOR_TYPE, LegacyPredefineIndicatorSchemaType, PredefineIndicatorV2SchemaType } from "@/entities/Indicator";
+import { LegacyPredefineIndicatorSchemaType, PredefineIndicatorV2SchemaType } from "@/entities/Indicator";
 import { INDICATORS_CATEGORIES } from "@/entities/IndicatorCategories";
+import { INDICATOR_TYPE } from "@/entities/IndicatorType";
+import { INDICATOR_ORDER } from "@/entities/IndicatorOrder";
 
 // TOXIC used in diaryData for old users, is registered as STATIC_UUID_FOR_INSTANCE_OF_GENERIC_INDICATOR_SUBSTANCE in
 // futur entries
@@ -1885,7 +1887,7 @@ export const INDICATEURS_LISTE_PAR_CATEGORIE = {
   ],
 };
 
-export const INDICATEURS_LISTE_ONBOARDING = [
+const INDICATEURS_LISTE_ONBOARDING = [
   "Humeur générale",
   "Anxiété",
   "Colère",
@@ -1912,7 +1914,7 @@ export const OLD_INDICATEURS_LISTE_ONBOARDING_SOMMEIL = [
   "Réveils nocturnes",
 ];
 
-export const INDICATEURS_LISTE_ONBOARDING_CUSTOM_SIMPLE_LIST = [
+const INDICATEURS_LISTE_ONBOARDING_CUSTOM_SIMPLE_LIST = [
   "Anxiété",
   "Confiance",
   "Colère",